Further version number adjustements
Done:

- Replace version number in setup.py with __version__.
- Add a number after 'dev'.

It is probably not the safest way. But as empymod has only few
dependencies, it hopefully won't create problems.

Adding git-version to dev-strings might still be a good idea for the
future. However, from the looks there is no clean/easy implementation,
not worth the hassle for me at the moment.