Fix mkl=2017.0.4 temporarily
In the recent anaconda with mkl-2018 there is a regression bug so that
numexpr is not built against VML. This results that numexpr will return
nan for "exp(-infty)", whereas numpy returns 0 for np.exp(-np.infty).
This causes certain tests to fail, which compare the normal versus the
numexpr version.

This fix is temporarily and should be undone when the regression bug is
resolved.

This issue should be resolved properly on the empymod-side. Currently
tests fail if the libraries are not built with mkl (using for instance
anaconda-forge instead of anaconda).