`numexpr` only used if `numexpr.use_vml=True`
There are various issues with `empymod` and `numexpr` if `numexpr` is
not built against Intel's VML. Therefore, `utils` checks if
`numexpr.use_vml` is  true, and disables the parallel option otherwise.
Fixes #3.

The main issue is that `evaluate("exp(+/-inf +/-infj)")` equals to
`+/-0 +/-0j` with `numpy` and `numexpr` with VML, but to `+/-nan +/-nanj`
for `numexpr` without VML. This causes issues for various scenarious in
various parts in the code.