Unify DLF arguments
Fixes #10.

These changes are backwards compatible for all main modelling routines
in `empymod.model`. However, they are not backwards compatible for the
following routines:
- `empymod.model.fem` (removed `use_spline`),
- `empymod.transform.fht` (removed `use_spline`),
- `empymod.transform.hqwe` (removed `use_spline`),
- `empymod.transform.quad` (removed `use_spline`),
- `empymod.transform.dlf` (`lagged`, `splined` repl. by `pts_per_dec`),
- `empymod.utils.check_opt` (no longer returns `use_spline`),
- `empymod.utils.check_hankel` (changes in `pts_per_dec`), and
- `empymod.utils.check_time` (changes in `pts_per_dec`).

The function `empymod.utils.spline_backwards_hankel` can be used for
backwards compatibility.

Now the Hankel and Fourier DLF have the same behaviour for
`pts_per_dec`:
- `pts_per_dec = 0`: Standard DLF,
- `pts_per_dec < 0`: Lagged Convolution DLF, and
- `pts_per_dec > 0`: Splined DLF.

This commit fixes the issue. Further commits are required to implement
properly the DeprecatedWarning as well as to add more testes.