Small change in check if kernels are emtpy
Small change (for speed) in check if kernels are empty in
``transform.dlf` and ``transform.qwe``: only roughly 100 values are
checked instead of the whole matrix. This has no influence on small
matrices, but can speed up the np.any()-call for very big matrices. I
don't think there is the chance that these 100 values are exactly zero
when in fact the kernel is not. However, something to keep an eye on.