`wavenumber` returns None instead of zeros
Kernels which do not exist for a given `ab` are now returned as `None`
from `kernel.wavenumber` instead of arrays of zeroes. This permits for
some time saving for the cases where one or two of the three kernels do
not exist. This change is backwards incompatible if you directly used
`kernel.wavenumber`. Nothing changes for the user-facing routines in
`model`.