np.log10 -> np.log
The benchmarks (https://empymod.github.io/asv) have revealed that every
time `log` was replaced by `log10`, there was a regression (9a0f3861,
575451e8, 90aa868a).

This is a try, replacing `log10` by `log` everywhere I see fit.

I will watch carefully the benchmarks for a decision to keep it this way
or not.